+++ src/main/scala/com/ironcorelabs/GroupAccessEditErr.scala
@@ -1,0 +1,12 @@
+package com.ironcorelabs.scala.sdk
+
+import com.ironcorelabs.{sdk => jsdk}
+
+case class GroupAccessEditErr(
+  user: UserId,
+  error: String
+)
+
+object GroupAccessEditErr {
+  def apply(gaee: jsdk.GroupAccessEditErr): GroupAccessEditErr = GroupAccessEditErr(UserId(gaee.getUser), gaee.getError)
+}