+++ src/main/scala/com/ironcorelabs/GroupAccessEditResult.scala
@@ -1,0 +1,16 @@
+package com.ironcorelabs.scala.sdk
+
+import com.ironcorelabs.{sdk => jsdk}
+
+case class GroupAccessEditResult(
+  succeeded: List[UserId],
+  failed: List[GroupAccessEditErr]
+)
+
+object GroupAccessEditResult {
+  def apply(gaer: jsdk.GroupAccessEditResult): GroupAccessEditResult =
+    GroupAccessEditResult(
+      gaer.getSucceeded.map(UserId(_)).toList,
+      gaer.getFailed.map(GroupAccessEditErr(_)).toList
+    )
+}