+++ src/main/scala/com/ironcorelabs/IronSdkFuture.scala
@@ -10,6 +10,14 @@ case class IronSdkFuture(deviceContext: DeviceContext) extends IronSdk[Future] {
   def groupCreate(options: GroupCreateOpts): Future[GroupMetaResult] =
     underlying.groupCreate(options).unsafeToFuture
 
+  def groupAddMembers(id: GroupId, users: List[UserId]): Future[GroupAccessEditResult] =
+    underlying.groupAddMembers(id, users).unsafeToFuture
+
+  def groupRemoveMembers(id: GroupId, userRevokes: List[UserId]): Future[GroupAccessEditResult] =
+    underlying.groupRemoveMembers(id, userRevokes).unsafeToFuture
+
+  def groupGetMetadata(id: GroupId): Future[GroupGetResult] = underlying.groupGetMetadata(id).unsafeToFuture
+
   def documentEncrypt(data: ByteVector, options: DocumentEncryptOpts): Future[DocumentEncryptResult] =
     underlying.documentEncrypt(data, options).unsafeToFuture